+++ config/symbols.pspeu.dra.txt
@@ -50,11 +50,25 @@ SetNextRoomToLoad = 0x090E73D0;
 func_800F0CD8 = 0x090E74E8;
 func_800F14CC = 0x090E7BC0;
 func_800F16D0 = 0x090E7DA8;
+func_800F1770 = 0x090E7E38;
+func_800F17C8 = 0x090E7E90;
+func_800F180C = 0x090E7ED8;
+func_800F1868 = 0x090E7F38;
+func_800F18C4 = 0x090E7F98;
+func_800F1954 = 0x090E8008;
+func_800F1A3C = 0x090E80E0;
+func_800F1B08 = 0x090E81A8;
+DrawSecretPassageOnMap = 0x090E83E8;
+RevealSecretPassageOnMap = 0x090E8518;
+RevealSecretPassageAtPlayerPositionOnMap = 0x090E8620;
 func_800F2014 = 0x090E8658;
+func_800F2120 = 0x090E8760;
 func_800F223C = 0x090E8880;
+func_800F2288 = 0x090E88C0;
 func_800F2404 = 0x090E8A30;
 func_800F24F4 = 0x090E8B30;
 DrawMapCursor = 0x090E8CA8;
+func_800F27F4 = 0x090E8E58;
 func_800F2860 = 0x090E8EC0;
 RunMainEngine = 0x090e9018;
 CheckIfAllButtonsAreAssigned = 0x090EB028;
@@ -227,7 +241,7 @@ func_800EB720 = 0x09104878;
 func_801092E8 = 0x09104958;
 func_80109328 = 0x091049E8;
 func_801093C4 = 0x09104AA0;
-func_80109594 = 0x09104C80;
+AluInit = 0x09104C80;
 func_80109990 = 0x09105200;
 CheckStageCollision = 0x091052B8;
 func_8010A234 = 0x09105DA0;
@@ -550,7 +564,11 @@ func_801361F8 = 0x09144BD8;
 CheckCollision = 0x09144C90;
 D_800A34C0 = 0x09145BD0;
 g_StagesLba = 0x091463D8;
+D_800A245C = 0x091474C0;
 D_800A297C = 0x091479E0;
+D_800A2D10 = 0x09147DC8;
+D_800A2D14 = 0x09147DF8;
+D_800A2D18 = 0x09147E28;
 D_800A2D58 = 0x09147E58;
 g_ChButtons = 0x0914A3A0;
 D_800A2D80 = 0x0914A3E0;
@@ -775,6 +793,7 @@ D_8006BB00 = 0x091CE5A8;
 g_DebugPlayer = 0x091CE5C0;
 g_LoadOvlIdx = 0x091CE5D0;
 D_80097420 = 0x091CE620;
+D_80097450 = 0x091CE650;
 g_PrevScrollY = 0x091CE6A0;
 g_PrevScrollX = 0x091CE6A8;
 g_RoomCount = 0x091CE6B8;
@@ -789,6 +808,7 @@ g_GfxLoad = 0x091ED538;
 D_8006C3C4 = 0x091F55F8;
 g_DemoMode = 0x091fc420;
 D_800973EC = 0x091fc428;
+g_MapCursorTimer = 0x091FC458;
 D_800978C4 = 0x091fc478;
 D_8006C378 = 0x091fc488;
 D_80097924 = 0x091FC490;
