+++ config/symbols.pspeu.txt
@@ -127,6 +127,7 @@ g_PlOvlSpritesheet = 0x08CE9C60;
 g_ServantDesc = 0x08D1DC40;
 D_8017A000 = 0x08D27C40;
 D_8017D000 = 0x08D2AC40;
+g_BmpCastleMap = 0x08D8DC40;
 DestroyEntitiesFromIndex = 0x0913E838;
 g_PauseAllowed = 0x091CE580;
 D_8003C744 = 0x091CE588;
