+++ config/symbols.us.bobo4.txt
@@ -139,7 +139,6 @@ EntitySoulStealOrb = 0x801C0DD8;
 EntityEnemyBlood = 0x801C116C;
 EntityRoomForeground = 0x801C161C;
 BottomCornerText = 0x801C1708;
-EntityDoppleganger10 = 0x801C2284;
 SetPlayerStep = 0x801C4958;
 func_8010D59C = 0x801C4970;
 func_8010D800 = 0x801C4BDC;
@@ -166,19 +165,14 @@ func_80111CC0 = 0x801C65F0;
 PlayerStepWalk = 0x801C6884;
 PlayerStepHighJump = 0x801C6F74;
 func_80113E68 = 0x801C7244;
-DopplegangerHandleDamage = 0x801C7380;
-PlayerStepKill = 0x801C78DC;
 BatFormFinished = 0x801C80D8;
 func_8011690C = 0x801C8184;
-ControlBatForm = 0x801C8224;
 PlayerStepSwordWarp = 0x801C95E4;
 GetFreeEntity = 0x801C9A6C;
 GetFreeEntityReverse = 0x801C9AE0;
 func_80118C28 = 0x801C9B64;
 DopEntityHitByHoly = 0x801C9CB8;
-PrintDummySet = 0x801C9FEC;
 CreateEntFactoryFromEntity = 0x801CA208;
-EntitySmokePuff = 0x801CA834;
 DopEntityHitByDark = 0x801CAE48;
 func_8011BD48 = 0x801CB020;
 EntityPlayerBlinkWhite = 0x801CB07C;
