+++ config/symbols.us.borbo5.txt
@@ -120,7 +120,6 @@ EntitySoulStealOrb = 0x801BFD0C;
 EntityEnemyBlood = 0x801C00A0;
 EntityRoomForeground = 0x801C0550;
 BottomCornerText = 0x801C063C;
-EntityDoppleganger40 = 0x801C11D4;
 MemCardSetPort = 0x801C1DB0;
 SetPlayerStep = 0x801C3F58;
 func_8010D59C = 0x801C3F70;
@@ -148,13 +147,9 @@ func_80111CC0 = 0x801C5C00;
 PlayerStepWalk = 0x801C5E94;
 PlayerStepHighJump = 0x801C6584;
 func_80113E68 = 0x801C6854;
-DopplegangerHandleDamage = 0x801C6990;
-PlayerStepKill = 0x801C6EEC;
 BatFormFinished = 0x801C76E8;
 func_8011690C = 0x801C7794;
-ControlBatForm = 0x801C7834;
 PlayerStepSwordWarp = 0x801C8BF4;
-PlayerStepStone = 0x801C8CA4;
 GetFreeEntity = 0x801C907C;
 GetFreeEntityReverse = 0x801C90F0;
 func_80118C28 = 0x801C9174;
@@ -162,7 +157,6 @@ func_80119E78 = 0x801C91D0;
 DopEntityHitByHoly = 0x801C92C8;
 PrintDummySet = 0x801C95FC;
 CreateEntFactoryFromEntity = 0x801C9818;
-EntitySmokePuff = 0x801C9E44;
 DopEntityHitByDark = 0x801CA458;
 func_8011BD48 = 0x801CA630;
 EntityPlayerOutline = 0x801CBCB4;
